+++ process_activity.go
@@ -14,11 +14,11 @@ import (
 
 type ProcessActivity struct {
 	Activity             string        `json:"activity_name" validate:"omitempty"`
-	ActivityID           uint8         `json:"activity_id" validate:"omitempty"`
+	ActivityID           int           `json:"activity_id" validate:"omitempty"`
 	Actor                *Actor        `json:"actor" validate:"omitempty"`
 	ActualPermissions    int           `json:"actual_permissions" validate:"omitempty"`
 	Category             string        `json:"category_name" validate:"omitempty"`
-	CategoryID           *uint8        `json:"category_uid" validate:"required"`
+	CategoryID           *int          `json:"category_uid" validate:"required"`
 	Class                string        `json:"class_name" validate:"omitempty"`
 	ClassID              *uint         `json:"class_uid" validate:"required"`
 	Count                uint          `json:"count" validate:"omitempty"`
@@ -38,12 +38,12 @@ type ProcessActivity struct {
 	RawData              string        `json:"raw_data" validate:"omitempty"`
 	RequestedPermissions int           `json:"requested_permissions" validate:"omitempty"`
 	Severity             string        `json:"severity" validate:"omitempty"`
-	SeverityID           uint8         `json:"severity_id" validate:"omitempty"`
+	SeverityID           int           `json:"severity_id" validate:"omitempty"`
 	StartTime            string        `json:"start_time" validate:"omitempty"`
 	Status               string        `json:"status" validate:"omitempty"`
 	StatusCode           string        `json:"status_code" validate:"omitempty"`
 	StatusDetails        string        `json:"status_detail" validate:"omitempty"`
-	StatusID             uint8         `json:"status_id" validate:"omitempty"`
+	StatusID             int           `json:"status_id" validate:"omitempty"`
 	TimezoneOffset       int           `json:"timezone_offset" validate:"omitempty"`
 	TypeID               *uint         `json:"type_uid" validate:"required"`
 	TypeName             string        `json:"type_name" validate:"omitempty"`
